requirements.txt
absl-py==2.3.0
astunparse==1.6.3
beautifulsoup4==4.12.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
Flask==2.3.3
flatbuffers==25.2.10
fonttools==4.58.2
gast==0.6.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.73.0
gunicorn==21.2.0
h5py==3.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==2.15.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.7.5
ml-dtypes==0.2.0
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
pillow==10.3.0
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
soupsieve==2.7
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==2.3.7
wrapt==1.14.1
